Visual Studio 2019 Cleanup
* Updated project and solution files to Visual Studio 2019
* Re-Targeted projects to Platform Toolset v142
* Adapted code to eliminate new warnings
* Suppress warnings from googletest and googlemock headers

+++ src/game_loop.cpp
@@ -94,7 +94,7 @@ void GameLoop::game_loop()
 		}
 		else {
 			tick++;
-			next_logic = t0 + (tick + 1) * freq / TPS;
+			next_logic = t0 + ((Uint64)tick + 1) * freq / TPS;
 		}
 	}
